+++ tests/conftest.py
@@ -1,18 +1,10 @@
 from functools import wraps
 from typing import List
 
-from docker import DockerClient
 from docker.models.containers import Container
 from pytest import fixture
 
-from yellowbox import docker_client as _docker_client
 from yellowbox.containers import create_and_pull as _create_and_pull
-
-
-@fixture(scope="module")
-def docker_client() -> DockerClient:
-    with _docker_client() as client:
-        yield client
 
 
 @fixture