+++ tests/extras/test_postgres.py
@@ -33,6 +33,27 @@ def test_local_connection(docker_client):
             assert vals == [2, 3]
 
 
+@mark.asyncio
+async def test_local_connection_async(docker_client):
+    service: PostgreSQLService
+    async with PostgreSQLService.arun(docker_client) as service:
+        with service.connection() as connection:
+            connection.execute("""
+            CREATE TABLE foo (x INTEGER, y TEXT);
+            INSERT INTO foo VALUES (1,'one'), (2, 'two'), (3, 'three'), (10, 'ten');
+            """)
+            connection.execute("""
+            DELETE FROM foo WHERE x = 10;
+            """)
+
+        with service.connection() as connection:
+            results = connection.execute("""
+            SELECT x, y FROM foo WHERE y like 't%%'
+            """)
+            vals = [row['x'] for row in results]
+            assert vals == [2, 3]
+
+
 def test_sibling(docker_client, create_and_pull):
     service: PostgreSQLService
     with PostgreSQLService.run(docker_client) as service:
@@ -49,7 +70,6 @@ def test_sibling(docker_client, create_and_pull):
             " -c 'DELETE FROM foo WHERE x < 3'",
             environment={'PGPASSWORD': service.password},
             detach=True,
-            remove=False,
         )
         container.start()
         return_status = container.wait()
@@ -123,7 +143,6 @@ def test_remote_connection_string(docker_client, create_and_pull):
             "python:latest",
             'sh -c "pip install sqlalchemy psycopg2 && python ./main.py"',
             detach=True,
-            remove=False  # todo
         )
         upload_file(
             container, './main.py',