+++ yellowbox/__init__.py
@@ -1,17 +1,20 @@
 from yellowbox._version import __version__
-from yellowbox.clients import docker_client
+from yellowbox.clients import docker_client, open_docker_client
 from yellowbox.image_build import build_image
 from yellowbox.networks import connect, temp_network
 from yellowbox.retry import RetrySpec
 from yellowbox.service import YellowService
-from yellowbox.subclasses import ContainerService, RunMixin, SingleContainerService, SingleEndpointService
+from yellowbox.subclasses import (
+    AsyncRunMixin, ContainerService, RunMixin, SingleContainerService, SingleEndpointService
+)
 
 __all__ = [
     '__version__',
     'docker_client',
+    'open_docker_client',
     'build_image',
     'connect', 'temp_network',
     'RetrySpec',
     'YellowService',
-    'ContainerService', 'RunMixin', 'SingleContainerService', 'SingleEndpointService',
+    'ContainerService', 'RunMixin', 'AsyncRunMixin', 'SingleContainerService', 'SingleEndpointService',
 ]