+++ yellowbox/subclasses.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
-from contextlib import contextmanager, nullcontext
-from typing import ContextManager, Iterator, Optional, Sequence, Type, TypeVar
+from contextlib import asynccontextmanager, contextmanager, nullcontext
+from typing import AsyncIterator, ContextManager, Iterator, Optional, Sequence, Type, TypeVar
 
 from docker import DockerClient
 from docker.models.containers import Container
@@ -10,7 +10,7 @@ import yellowbox.networks as networks_mod
 from yellowbox.containers import _DEFAULT_TIMEOUT, get_aliases, is_alive, is_removed
 from yellowbox.retry import RetrySpec
 from yellowbox.service import YellowService
-from yellowbox.utils import _get_spinner
+from yellowbox.utils import _SPINNER_SUCCESSMSG, _get_spinner
 
 
 class ContainerService(YellowService):
@@ -129,8 +129,6 @@ class RunMixin:
             retry_spec: Optional[RetrySpec] = None,
             network: Optional[Network] = None, **kwargs) -> Iterator[_T]:
         """
-        Same as RunMixin.run, but allows to forward retry arguments to the blocking start method.
-
         Args:
             docker_client: a DockerClient instance to use when creating the service
             spinner: whether or not to use a yaspin spinner
@@ -151,5 +149,54 @@ class RunMixin:
         with connect_network:
             with yaspin_spinner(f"Waiting for {cls.service_name()} to start..."):  # type: ignore[attr-defined]
                 service.start(retry_spec=retry_spec)
+            with service:
+                yield service
+
+
+class AsyncRunMixin(ABC):
+    @classmethod
+    def service_name(cls):
+        return cls.__name__
+
+    @abstractmethod
+    async def astart(self: _T, retry_spec: Optional[RetrySpec] = None) -> None:   # type: ignore[misc]
+        """
+        Start the service synchronously, but block and wait for startup asynchronously.
+        """
+        pass
+
+    @classmethod
+    @asynccontextmanager
+    async def arun(cls: Type[_T], docker_client: DockerClient, *, verbose: bool = True,  # type: ignore[misc]
+                   retry_spec: Optional[RetrySpec] = None,
+                   network: Optional[Network] = None, **kwargs) -> AsyncIterator[_T]:
+        """
+        Same as RunMixin.run, but waits for startup asynchronously.
+
+        Args:
+            docker_client: a DockerClient instance to use when creating the service
+            verbose: whether or not to print progress information when setting up the service
+            retry_spec: forwarded to cls.start
+            network: connect service to network
+            **kwargs: all keyword arguments are forwarded to the class's constructor
+        """
+
+        yaspin_spinner = _get_spinner(verbose)
+        with yaspin_spinner(f"Fetching {cls.service_name()} ..."):  # type: ignore[attr-defined]
+            service = cls(docker_client, **kwargs)
+
+        connect_network: ContextManager[None]
+        if network:
+            connect_network = networks_mod.connect(network, service)
+        else:
+            connect_network = nullcontext()
+
+        with connect_network:
+            if verbose:
+                print(f".   Waiting for {cls.service_name()} to start...")    # type: ignore[attr-defined]
+            await service.astart(retry_spec=retry_spec)    # type: ignore[attr-defined]
+            if verbose:
+                print(f"{_SPINNER_SUCCESSMSG} "    # type: ignore[attr-defined]
+                      f"{cls.service_name()} started successfully")    # type: ignore[attr-defined]
             with service:
                 yield service